refactor estados
