--- rpi5/preprocessing/fft.py
@@ -1,19 +1,0 @@
-from numpy.fft import fft2, fftshift
-
-def ideal_lowpass(I, factor=0.6, method='Stone_et_al_2001'):
-    if method == 'Stone_et_al_2001':
-        m = factor * I.shape[0]/2
-        n = factor * I.shape[1]/2
-        N = np.min([m, n])
-        I = I[int(I.shape[0] // 2 - N): int(I.shape[0] // 2 + N),
-            int(I.shape[1] // 2 - N): int(I.shape[1] // 2 + N)]
-        return I
-    else:
-        raise ValueError('Método não suportado.')
-
-
-def image_preprocessing(image, method='Stone_et_al_2001'):
-    fft_from_image = fftshift(fft2(image))
-    if method is not None:
-        fft_from_image = ideal_lowpass(fft_from_image, method=method)
-    return fft_from_image
