--- rpi5/preprocessing/grayscale.py
@@ -1,10 +1,0 @@
-import numpy as np
-import cv2
-from PIL import Image, ImageOps
-
-def cv2_to_nparray_grayscale(frame):
-    cv2_img = cv2.cvtColor(frame, cv2.COLOR_BGR2RGB)
-    img_array_rgb = Image.fromarray(cv2_img)
-    img_grayscale = ImageOps.grayscale(img_array_rgb)
-    img_array = np.asarray(img_grayscale)
-    return img_array
