--- rpi5/preprocessing/window.py
@@ -1,33 +1,0 @@
-import numpy as np
-
-def apply_blackman_harris_window(image):
-    height, width = image.shape
-    window_row = blackman_harris_window(width)
-    window_col = blackman_harris_window(height)
-    image_windowed = np.outer(window_col, window_row) * image
-    return image_windowed
-
-def blackman_harris_window(size, a0=0.35875, a1=0.48829, a2=0.14128, a3=0.01168):
-    # a0, a1, a2 e a3 são os coeficientes de janelamento
-    # Criação do vetor de amostras
-    n = np.arange(size)
-    # Cálculo da janela de Blackman-Harris
-    window = a0 - a1 * np.cos(2 * np.pi * n / (size - 1)) + a2 * np.cos(4 * np.pi * n / (size - 1)) - a3 * np.cos(
-        6 * np.pi * n / (size - 1))
-    return window
-
-def apply_raised_cosine_window(image):
-    rows, cols = image.shape
-    i = np.arange(rows)
-    j = np.arange(cols)
-    window = 0.5 * (1 + np.cos(np.pi * (2 * i[:, None] - rows) / rows)) * \
-             0.5 * (1 + np.cos(np.pi * (2 * j - cols) / cols))
-    return image * window
-
-def apply_border_windowing_on_image(image, border_windowing_method="blackman_harris"):
-    if border_windowing_method == "blackman_harris":
-        return apply_blackman_harris_window(image)
-    elif border_windowing_method == "raised_cosine":
-        return apply_raised_cosine_window(image)
-    elif border_windowing_method == None:
-        return image
